Support partial share purchases/sales
Something that is possible through some brokerages is the purchase and
sale of partial shares. This means that the `qty` value isn't
necessarily an integer, but could be represented as a floating point.
This can also occur for things like bitcoin or currency, where you will
most likely hold a certain fraction of a coin/currency. The primary
change was to support having decimal/float values for the quantity.

Additionally, I noticed that we were doing a lot of floating point math.
While this is good for approximations, it is better to be precise. I
discovered that there is a `decimal` python library that allows you to
work with precise decimal objects that don't have weird rounding rules
like floats. Anywhere where we did floating point math or where we
returned floats, I converted to use a Decimal instead.

+++ capgains/exchange_rate.py
@@ -1,5 +1,6 @@
 import requests
 from datetime import date, timedelta, datetime
+from decimal import Decimal
 from click import ClickException
 
 
@@ -83,7 +84,7 @@ class ExchangeRate:
         for day_rate in rates_json:
             date_str = day_rate[self.date]
             date = datetime.strptime(date_str, '%Y-%m-%d').date()
-            rate = float(day_rate[forex_str][self.value])
+            rate = Decimal(day_rate[forex_str][self.value])
             rates[date] = rate
         return rates
 
@@ -130,7 +131,7 @@ class ExchangeRate:
             not exist for that day"""
         if self._currency_from == self.currency_to:
             # Converting CAD to CAD
-            rate = 1.00
+            rate = Decimal(1.00)
         else:
             # Converting Non-CAD to CAD
             rate = self._get_closest_rate_for_day(date)
